Add explicit trailing slash when saving question
If the trailing slash is omitted from the `NOTESDIR` declaration,
QuickQuestion behaves a little unexpectedly. A note is written one
directory up from where you expect, with the last `NOTESDIR` component
included in the prefix.

For example, before the change:

    NOTESDIR="/path/to/some/dir"
    NOTESEXT="txt"
    NOTESPRE="qq"
    # ...
    qq -a "question" "answer"
    ls /path/to/some
    # dirqq question.txt

After the change:

    NOTESDIR="/path/to/some/dir"
    NOTESEXT="txt"
    NOTESPRE="qq"
    # ...
    qq -a "question" "answer"
    ls /path/to/some/dir
    # qq question.txt

If `NOTESDIR` itself includes a trailing slash, everything still works
as before.

Admittedly this for saving the user from himself.